代码迁移重构

+++ src/main/java/com/gameplat/admin/model/bean/ProxyPayMerBean.java
@@ -3,19 +3,21 @@ package com.gameplat.admin.model.bean;
 import com.alibaba.fastjson.JSON;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 /** 转换类 */
 @Data
 public class ProxyPayMerBean {
 
-  private Long maxLimitCash; // 最大金额限制
+  private BigDecimal maxLimitCash; // 最大金额限制
 
-  private Long minLimitCash; // 最小金额限制
+  private BigDecimal minLimitCash; // 最小金额限制
 
   private String userLever; // 用户层级
 
   public ProxyPayMerBean() {}
 
-  public ProxyPayMerBean(Long maxLimitCash, Long minLimitCash, String userLever) {
+  public ProxyPayMerBean(BigDecimal maxLimitCash, BigDecimal minLimitCash, String userLever) {
     this.maxLimitCash = maxLimitCash;
     this.minLimitCash = minLimitCash;
     this.userLever = userLever;
