+++ src/main/java/com/gameplat/admin/model/vo/PpMerchantVO.java
@@ -3,6 +3,8 @@ package com.gameplat.admin.model.vo;
 import com.baomidou.mybatisplus.extension.activerecord.Model;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class PpMerchantVO extends Model<PpMerchantVO> {
 
@@ -18,7 +20,7 @@ public class PpMerchantVO extends Model<PpMerchantVO> {
 
   private Long proxyTimes;
 
-  private Long proxyAmount;
+  private BigDecimal proxyAmount;
 
   private String interfaceName;
 
@@ -26,9 +28,9 @@ public class PpMerchantVO extends Model<PpMerchantVO> {
 
   private String merLimits;
 
-  private Long maxLimitCash; // 最大金额限制
+  private BigDecimal maxLimitCash; // 最大金额限制
 
-  private Long minLimitCash; // 最小金额限制
+  private BigDecimal minLimitCash; // 最小金额限制
 
   private String userLever; // 用户层级
